backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
Flask-Mail==0.9.1
celery==5.3.4
redis==5.0.1